Refactor LlamaIndexService import paths and consolidate service files
