+++ lib/llamaindex-service.ts
